requirements.txt
APScheduler==3.6.3
beautifulsoup4==4.10.0
certifi==2021.10.8
chardet==3.0.4
idna==2.10
python-dotenv==0.12.0
python-telegram-bot==13.4.1
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.23.0
six==1.16.0
soupsieve==2.3.1
tabulate==0.8.9
tornado==6.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.25.11
